functions/requirements.txt
python-telegram-bot
llama-cpp-python
functions-framework
google-cloud-pubsub
google-cloud-firestore
google-cloud-storage
google-cloud-logging
Flask
PyYAML